third_party/WebKit/LayoutTests/imagecapture/getPhotoSettings.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/mock-imagecapture.js"></script>
<body>
<canvas id='canvas' width=10 height=10/>
</body>
<script>

const fillLightModeNames = ['off', 'auto', 'flash'];

// This test verifies that ImageCapture can call getPhotoSettings(), with a
// mock Mojo interface implementation.
async_test(function(t) {
  var canvas = document.getElementById('canvas');
  var context = canvas.getContext('2d');
  context.fillStyle = 'red';
  context.fillRect(0, 0, 10, 10);
  var stream = canvas.captureStream();

  var mock_state;

  mockImageCaptureReady
    .then(mock => {
      mock_state = mock.state();
      return new ImageCapture(stream.getVideoTracks()[0]);
    },
    error => {
      assert_unreached('Error creating MockImageCapture: ' + error);
    })
    .then(capturer => {
      return capturer.getPhotoSettings();
    })
    .then(settings => {
      assert_equals(settings.imageWidth, mock_state.width.current, 'width');
      assert_equals(settings.imageHeight, mock_state.height.current, 'height');
      // TODO(mcasas): check the remaining two entries https://crbug.com/732521.

      t.done();
    })
    .catch(error => {
      assert_unreached('Error during getPhotoSettings(): ' + error.message);
    });
}, 'exercises ImageCapture.getPhotoSettings()');

</script>
